+++ packages/router/src/adapters/subgraph/subgraph.ts
@@ -13,12 +13,7 @@ import { BigNumber, constants } from "ethers/lib/ethers";
 
 import { getContext } from "../../router";
 import { ContractReaderNotAvailableForChain, NoChainConfig } from "../../lib/errors";
-import {
-  ActiveTransaction,
-  SingleChainTransaction,
-  CrosschainTransactionStatus,
-  CancelPayload,
-} from "../../lib/entities";
+import { ActiveTransaction, SingleChainTransaction, CrosschainTransactionStatus } from "../../lib/entities";
 import { handlingTracker } from "../../bindings/contractReader";
 
 import {
@@ -65,7 +60,7 @@ const setSyncRecord = async (chainId: number, requestContext: RequestContext): P
       latestBlock,
       records: records.map((r) => ({ synced: r.synced, lag: r.lag, syncedBlock: r.syncedBlock, uri: r.uri })),
     });
-  } catch (e) {
+  } catch (e: any) {
     logger.error(`Error getting sync records for chain ${chainId}`, requestContext, methodContext, jsonifyError(e), {
       chainId,
     });
@@ -171,7 +166,16 @@ export const getActiveTransactions = async (_requestContext?: RequestContext): P
         const receiverNotConfigured = toCancel.map((senderTx) => {
           return {
             crosschainTx: sdkSenderTransactionToCrosschainTransaction(senderTx),
-            payload: {},
+            payload: {
+              hashes: {
+                sending: {
+                  prepareHash: senderTx.prepareTransactionHash,
+                  cancelHash: senderTx.cancelTransactionHash,
+                  fulfillHash: senderTx.fulfillTransactionHash,
+                },
+                receiving: {},
+              },
+            },
             status: CrosschainTransactionStatus.ReceiverNotConfigured,
           } as ActiveTransaction<"ReceiverNotConfigured">;
         });
@@ -242,7 +246,22 @@ export const getActiveTransactions = async (_requestContext?: RequestContext): P
                   sending,
                   receiving,
                 },
-                payload: {},
+                payload: {
+                  hashes: {
+                    sending: {
+                      prepareHash: senderTx.prepareTransactionHash,
+                      cancelHash: senderTx.cancelTransactionHash,
+                      fulfillHash: senderTx.fulfillTransactionHash,
+                    },
+                    receiving: correspondingReceiverTx
+                      ? {
+                          prepareHash: correspondingReceiverTx.prepareTransactionHash,
+                          cancelHash: correspondingReceiverTx.cancelTransactionHash,
+                          fulfillHash: correspondingReceiverTx.fulfillTransactionHash,
+                        }
+                      : undefined,
+                  },
+                },
                 status: CrosschainTransactionStatus.SenderExpired,
               } as ActiveTransaction<"SenderExpired">;
             }
@@ -253,7 +272,15 @@ export const getActiveTransactions = async (_requestContext?: RequestContext): P
               if (!receiverSynced) {
                 return {
                   crosschainTx: sdkSenderTransactionToCrosschainTransaction(senderTx),
-                  payload: {},
+                  payload: {
+                    hashes: {
+                      sending: {
+                        prepareHash: senderTx.prepareTransactionHash,
+                        cancelHash: senderTx.cancelTransactionHash,
+                        fulfillHash: senderTx.fulfillTransactionHash,
+                      },
+                    },
+                  },
                   status: CrosschainTransactionStatus.ReceiverNotConfigured,
                 } as ActiveTransaction<"ReceiverNotConfigured">;
               }
@@ -268,6 +295,13 @@ export const getActiveTransactions = async (_requestContext?: RequestContext): P
                   encodedBid: senderTx.encodedBid,
                   encryptedCallData: senderTx.encryptedCallData,
                   senderPreparedHash: senderTx.prepareTransactionHash,
+                  hashes: {
+                    sending: {
+                      prepareHash: senderTx.prepareTransactionHash,
+                      cancelHash: senderTx.cancelTransactionHash,
+                      fulfillHash: senderTx.fulfillTransactionHash,
+                    },
+                  },
                 },
                 status: CrosschainTransactionStatus.SenderPrepared,
               } as ActiveTransaction<"SenderPrepared">;
@@ -299,7 +333,20 @@ export const getActiveTransactions = async (_requestContext?: RequestContext): P
                   signature: correspondingReceiverTx?.signature,
                   relayerFee: correspondingReceiverTx?.relayerFee,
                   callData: correspondingReceiverTx?.callData,
-                  receiverFulfilledHash: correspondingReceiverTx?.fulfillTransactionHash,
+                  hashes: {
+                    sending: {
+                      prepareHash: senderTx.prepareTransactionHash,
+                      cancelHash: senderTx.cancelTransactionHash,
+                      fulfillHash: senderTx.fulfillTransactionHash,
+                    },
+                    receiving: correspondingReceiverTx
+                      ? {
+                          prepareHash: correspondingReceiverTx.prepareTransactionHash,
+                          cancelHash: correspondingReceiverTx.cancelTransactionHash,
+                          fulfillHash: correspondingReceiverTx.fulfillTransactionHash,
+                        }
+                      : undefined,
+                  },
                 },
                 status: CrosschainTransactionStatus.ReceiverFulfilled,
               } as ActiveTransaction<"ReceiverFulfilled">;
@@ -312,7 +359,22 @@ export const getActiveTransactions = async (_requestContext?: RequestContext): P
                   sending,
                   receiving,
                 },
-                payload: {} as CancelPayload,
+                payload: {
+                  hashes: {
+                    sending: {
+                      prepareHash: senderTx.prepareTransactionHash,
+                      cancelHash: senderTx.cancelTransactionHash,
+                      fulfillHash: senderTx.fulfillTransactionHash,
+                    },
+                    receiving: correspondingReceiverTx
+                      ? {
+                          prepareHash: correspondingReceiverTx.prepareTransactionHash,
+                          cancelHash: correspondingReceiverTx.cancelTransactionHash,
+                          fulfillHash: correspondingReceiverTx.fulfillTransactionHash,
+                        }
+                      : undefined,
+                  },
+                },
                 status: CrosschainTransactionStatus.ReceiverCancelled,
               };
             }
@@ -337,12 +399,20 @@ export const getActiveTransactions = async (_requestContext?: RequestContext): P
                 invariant,
                 receiving,
               },
-              payload: {},
+              payload: {
+                hashes: {
+                  receiving: {
+                    prepareHash: expTx.prepareTransactionHash,
+                    cancelHash: expTx.cancelTransactionHash,
+                    fulfillHash: expTx.fulfillTransactionHash,
+                  },
+                },
+              },
               status: CrosschainTransactionStatus.ReceiverExpired,
             } as ActiveTransaction<"ReceiverExpired">;
           });
         return filterUndefined.concat(receiverNotConfigured).concat(remainingReceiverExpired);
-      } catch (e) {
+      } catch (e: any) {
         // Set this chain's error.
         errors.set(cId, e);
         logger.error(