+++ packages/router/src/bindings/contractReader/index.ts
@@ -62,7 +62,7 @@ export const bindContractReader = async () => {
           handlingTracker: [...handlingTracker],
         });
       }
-    } catch (err) {
+    } catch (err: any) {
       logger.error("Error getting active txs, waiting for next loop", requestContext, methodContext, jsonifyError(err));
       return;
     }
@@ -182,16 +182,22 @@ export const handleSingle = async (
         requestContext,
       });
     }
+    if (!_transaction.payload.hashes.sending) {
+      logger.warn("No sending hashes with SenderPrepared status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     const senderPrepareReceipt = await txService.getTransactionReceipt(
       _transaction.crosschainTx.invariant.sendingChainId,
-      _transaction.payload.senderPreparedHash,
+      _transaction.payload.hashes.sending.prepareHash,
     );
     if ((senderPrepareReceipt?.confirmations ?? 0) < chainConfig.confirmations) {
       logger.info("Waiting for safe confirmations", requestContext, methodContext, {
         txConfirmations: senderPrepareReceipt?.confirmations ?? 0,
         configuredConfirmations: chainConfig.confirmations,
         chainId: _transaction.crosschainTx.invariant.sendingChainId,
-        txHash: _transaction.payload.senderPreparedHash,
+        txHash: _transaction.payload.hashes.sending.prepareHash,
       });
       return;
     }
@@ -225,7 +231,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.receivingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const json = jsonifyError(err);
       if (safeJsonStringify(json).includes("#P:015")) {
         logger.warn("Receiver transaction already prepared", requestContext, methodContext, { error: json });
@@ -248,6 +254,7 @@ export const handleSingle = async (
               amount: transaction.crosschainTx.sending.amount,
               expiry: transaction.crosschainTx.sending.expiry,
               preparedBlockNumber: transaction.crosschainTx.sending.preparedBlockNumber,
+              preparedTransactionHash: transaction.payload.hashes.sending.prepareHash,
               side: "sender",
             },
             requestContext,
@@ -262,7 +269,7 @@ export const handleSingle = async (
             chainId: _transaction.crosschainTx.invariant.sendingChainId,
             amount: cancelRes!.gasUsed.toString(),
           });
-        } catch (cancelErr) {
+        } catch (cancelErr: any) {
           const cancelJson = jsonifyError(cancelErr);
           if (safeJsonStringify(jsonifyError(cancelErr)).includes("#C:019")) {
             logger.warn("Already cancelled", requestContext, methodContext, {
@@ -288,14 +295,20 @@ export const handleSingle = async (
       // this should not happen, this should get checked before this point
       throw new ContractReaderNotAvailableForChain(_transaction.crosschainTx.invariant.sendingChainId, {});
     }
+    if (!_transaction.payload.hashes.receiving) {
+      logger.warn("No receiving hashes with ReceiverFulfilled status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     const receiverPrepareReceipt = await txService.getTransactionReceipt(
       _transaction.crosschainTx.invariant.receivingChainId,
-      _transaction.payload.receiverFulfilledHash,
+      _transaction.payload.hashes.receiving.prepareHash,
     );
     if ((receiverPrepareReceipt?.confirmations ?? 0) < chainConfig.confirmations) {
       logger.info("Waiting for safe confirmations", requestContext, methodContext, {
         chainId: _transaction.crosschainTx.invariant.receivingChainId,
-        txHash: _transaction.payload.receiverFulfilledHash,
+        txHash: _transaction.payload.hashes.receiving.prepareHash,
         txConfirmations: receiverPrepareReceipt?.confirmations ?? 0,
         configuredConfirmations: chainConfig.confirmations,
       });
@@ -347,7 +360,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.sendingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const jsonErr = jsonifyError(err);
       if (safeJsonStringify(jsonErr).includes("#F:019")) {
         logger.warn("Sender already fulfilled", requestContext, methodContext, { error: jsonErr });
@@ -367,6 +380,12 @@ export const handleSingle = async (
       transaction.crosschainTx.invariant.transactionId,
     );
     const _transaction = transaction as ActiveTransaction<"ReceiverExpired">;
+    if (!_transaction.payload.hashes.receiving) {
+      logger.warn("No receiving hashes with ReceiverExpired status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     try {
       logger.info("Cancelling expired receiver", requestContext, methodContext);
       receipt = await cancel(
@@ -375,6 +394,7 @@ export const handleSingle = async (
           amount: _transaction.crosschainTx.receiving!.amount,
           expiry: _transaction.crosschainTx.receiving!.expiry,
           preparedBlockNumber: _transaction.crosschainTx.receiving!.preparedBlockNumber,
+          preparedTransactionHash: _transaction.payload.hashes.receiving.prepareHash,
           side: "receiver",
         },
         requestContext,
@@ -391,7 +411,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.receivingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const errJson = jsonifyError(err);
       if (safeJsonStringify(errJson).includes("#C:019")) {
         logger.warn("Already cancelled", requestContext, methodContext, { error: errJson });
@@ -411,6 +431,12 @@ export const handleSingle = async (
       transaction.crosschainTx.invariant.transactionId,
     );
     const _transaction = transaction as ActiveTransaction<"SenderExpired">;
+    if (!_transaction.payload.hashes.sending) {
+      logger.warn("No sending hashes with SenderExpired status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     try {
       logger.info("Cancelling expired sender", requestContext, methodContext);
       receipt = await cancel(
@@ -419,6 +445,7 @@ export const handleSingle = async (
           amount: _transaction.crosschainTx.sending.amount,
           expiry: _transaction.crosschainTx.sending.expiry,
           preparedBlockNumber: _transaction.crosschainTx.sending.preparedBlockNumber,
+          preparedTransactionHash: _transaction.payload.hashes.sending.prepareHash,
           side: "sender",
         },
         requestContext,
@@ -433,7 +460,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.sendingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const errJson = jsonifyError(err);
       if (safeJsonStringify(errJson).includes("#C:019")) {
         logger.warn("Already cancelled", requestContext, methodContext, { error: errJson });
@@ -457,6 +484,12 @@ export const handleSingle = async (
       "ContractReader => ReceiverCancelled",
       transaction.crosschainTx.invariant.transactionId,
     );
+    if (!_transaction.payload.hashes.sending) {
+      logger.warn("No sending hashes with ReceiverCancelled status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     try {
       logger.info("Cancelling sender after receiver cancelled", requestContext, methodContext);
       receipt = await cancel(
@@ -465,6 +498,7 @@ export const handleSingle = async (
           amount: _transaction.crosschainTx.sending.amount,
           expiry: _transaction.crosschainTx.sending.expiry,
           preparedBlockNumber: _transaction.crosschainTx.sending.preparedBlockNumber,
+          preparedTransactionHash: _transaction.payload.hashes.sending.prepareHash,
           side: "sender",
         },
         requestContext,
@@ -479,7 +513,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.sendingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const errJson = jsonifyError(err);
       if (safeJsonStringify(errJson).includes("#C:019")) {
         logger.warn("Already cancelled", requestContext, methodContext, { error: errJson });
@@ -500,6 +534,12 @@ export const handleSingle = async (
       "ContractReader => ReceiverNotConfigured",
       transaction.crosschainTx.invariant.transactionId,
     );
+    if (!_transaction.payload.hashes.sending) {
+      logger.warn("No sending hashes with ReceiverNotConfigured status", requestContext, methodContext, {
+        hashes: _transaction.payload.hashes,
+      });
+      return;
+    }
     try {
       logger.info("Cancelling sender because receiver is not configured", requestContext, methodContext);
       receipt = await cancel(
@@ -508,6 +548,7 @@ export const handleSingle = async (
           amount: _transaction.crosschainTx.sending.amount,
           expiry: _transaction.crosschainTx.sending.expiry,
           preparedBlockNumber: _transaction.crosschainTx.sending.preparedBlockNumber,
+          preparedTransactionHash: _transaction.payload.hashes.sending.prepareHash,
           side: "sender",
         },
         requestContext,
@@ -522,7 +563,7 @@ export const handleSingle = async (
         chainId: _transaction.crosschainTx.invariant.sendingChainId,
         amount: receipt!.gasUsed.toString(),
       });
-    } catch (err) {
+    } catch (err: any) {
       const errJson = jsonifyError(err);
       if (safeJsonStringify(errJson).includes("#C:019")) {
         logger.warn("Already cancelled", requestContext, methodContext, { error: errJson });